rename dl => platform
