+++ oneflow/core/platform/include/wrapper.h
@@ -13,8 +13,8 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-#ifndef ONEFLOW_CORE_DL_INCLUDE_WRAPPER_H_
-#define ONEFLOW_CORE_DL_INCLUDE_WRAPPER_H_
+#ifndef ONEFLOW_CORE_PLATFORM_INCLUDE_WRAPPER_H_
+#define ONEFLOW_CORE_PLATFORM_INCLUDE_WRAPPER_H_
 
 #include "oneflow/core/common/util.h"
 
@@ -40,4 +40,4 @@ class DynamicLibrary {
 
 }  // namespace oneflow
 
-#endif  // ONEFLOW_CORE_DL_INCLUDE_WRAPPER_H_
+#endif  // ONEFLOW_CORE_PLATFORM_INCLUDE_WRAPPER_H_
